quizzes/DiRAC-drivers-license: First pass at automating the exam
See:
  http://software-carpentry.org/blog/2012/08/alpha-test-of-drivers-license-exam.html
  http://software-carpentry.org/blog/2013/02/dirac-dry-run-two.html

I was unable to convert the code review question to the quizzer
format, because it's hard to evaluate refactorings automatically.

This quiz is probably not under the GPLv3.  The licensing for the
original quiz is unclear, but most Software Carpentry stuff is under
the Creative Commons Attribution 3.0 Unported (CC BY 3.0).